backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
python-multipart
pydantic-settings
stripe
httpx
python-dotenv
celery==5.3.6
redis==5.0.1
flower==2.0.1
email-validator
prometheus-fastapi-instrumentator
openai
pytest
pytest-asyncio
httpx
